+++ src/map.h
@@ -96,7 +96,7 @@ extern bool Map_IsValidPosition(uint16 position);
 extern bool Map_Save(FILE *fp);
 extern bool Map_Load(FILE *fp, uint32 length);
 extern bool Map_IsPositionUnveiled(uint16 position);
-extern bool Map_IsPositionInViewport(tile32 position, uint16 *retX, uint16 *retY);
+extern bool Map_IsPositionInViewport(tile32 position, int *retX, int *retY);
 extern void Map_MakeExplosion(uint16 type, tile32 position, uint16 hitpoints, uint16 unitOriginEncoded);
 extern uint16 Map_GetLandscapeType(uint16 packed);
 extern void Map_Update(uint16 packed, uint16 type, bool ignoreInvisible);
